+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Add the attribute `@coq_precise_signature` to help to distinguish between ambiguous signatures.
 * Remove the generation of `Import` in Coq and expand the names regardless of the `open` commands, in order to be compatible with local opens.
 * Add support for anonymous sub-signatures.
 * Generate functors using a type class of the current arguments.